requirements.txt
langchain
langchain_openai
langchain_community
streamlit
wikipedia
python-dotenv
numexpr